move readArgs
